refactor: combine subscribe using paypal and stripe
