+++ src/scenarios/billing/SubscribeToSubscriptionPlan/PaypalPaymentProvider/graphql-documents.ts
@@ -1,40 +1,11 @@
 import gql from 'graphql-tag';
 
-export const getPaymentPlansQuery = gql`
-  query GetPaymentPlans {
-    paymentPlans(
-      filter: {
-        providerConfigs: { some: { paymentProviderKey: { equalTo: "PAYPAL" } } }
-      }
-    ) {
-      nodes {
-        id
-        subscriptionPlan {
-          title
-        }
-        title
-        periodQuantity
-        periodUnit
-      }
-    }
-    paypalSettings(filter: { paymentProviderKey: { equalTo: "PAYPAL" } }) {
+export const getPaypalSettingsQuery = gql`
+  query GetPaypalSettings {
+    paypalSettings {
       nodes {
         clientId
         isSandbox
-      }
-    }
-  }
-`;
-
-export const getPaymentPlanPricesQuery = gql`
-  query GetPaymentPlanPrices($id: UUID!) {
-    paymentPlan(id: $id) {
-      prices {
-        nodes {
-          country
-          currency
-          price
-        }
       }
     }
   }
