+++ src/scenarios/billing/index.ts
@@ -1,8 +1,7 @@
-export * from './SubscribeToSubscriptionPlanPayPal/SubscribeToSubscriptionPlanPayPal';
 export * from './ListSubscriptionPlansAnonymously/ListSubscriptionPlansAnonymously';
 export * from './ListUserSubscriptions/ListUserSubscriptions';
 export * from './ListSubscriptionPlansForUser/ListSubscriptionPlansForUser';
 export * from './UnsubscribeFromSubscriptionPlanPayPal/UnsubscribeFromSubscriptionPlanPayPal';
 export * from './ListUserPaymentHistory/ListUserPaymentHistory';
-export * from './SubscribeToSubscriptionPlanStripe/SubscribeToSubscriptionPlanStripe';
 export * from './UnsubscribeFromSubscriptionPlanStripe/UnsubscribeFromSubscriptionPlanStripe';
+export * from './SubscribeToSubscriptionPlan/SubscribeToSubscriptionPlan';
